+++ tests/test_unit/test_cdf_tk/test_commands/test_migration_cmd/test_command.py
@@ -1343,21 +1343,6 @@ class TestMigrationCommand:
             client.data_modeling.statistics.project.return_value = stats
             cmd.validate_available_capacity(client, 10_000)
 
-    @pytest.mark.parametrize(
-        "existing_files, stem, expected",
-        [
-            ([], "migration_log", "migration_log-"),
-            (["migration_log-part001.log"], "migration_log", "migration_log-run2-"),
-            (["migration_log-part001.log", "migration_log-run3-part001.log"], "migration_log", "migration_log-run4-"),
-        ],
-    )
-    def test_create_logfile_stem(self, existing_files: list[str], stem: str, expected: str, tmp_path: Path) -> None:
-        for filename in existing_files:
-            (tmp_path / filename).touch()
-        actual = MigrationCommand._create_logfile_stem(tmp_path, stem, "not_important")
-
-        assert actual == expected
-
     def test_validate_stream_capacity_insufficient_records(self) -> None:
         stream = _make_stream_response(provisioned_records=1_000, consumed_records=900)
         with pytest.raises(ToolkitValueError, match="enough record capacity"):
